+++ app/src/main/java/br/digitalhouse/marveltime/FavoritosFragment.java
@@ -1,33 +1,19 @@
 package br.digitalhouse.marveltime;
-
-
 import android.os.Bundle;
-
 import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-
-/**
- * A simple {@link Fragment} subclass.
- */
 public class FavoritosFragment extends Fragment {
 
-
-    public FavoritosFragment() {
-        // Required empty public constructor
-    }
-
+    public FavoritosFragment() {  }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_favoritos, container, false);
-
-        return view;
+         View view = inflater.inflate(R.layout.fragment_favoritos, container, false);
+         return view;
     }
 
 }
