+++ app/src/main/res/layout/activity_cadastro.xml
@@ -11,11 +11,11 @@
         android:id="@+id/imgLogoMarvel"
         android:layout_width="264dp"
         android:layout_height="117dp"
-        android:layout_marginStart="64dp"
-        android:layout_marginLeft="64dp"
+        android:layout_marginStart="65dp"
+        android:layout_marginLeft="65dp"
         android:layout_marginTop="8dp"
-        android:layout_marginEnd="66dp"
-        android:layout_marginRight="66dp"
+        android:layout_marginEnd="65dp"
+        android:layout_marginRight="65dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
