+++ app/src/main/res/layout/activity_main.xml
@@ -8,17 +8,18 @@
     tools:context=".MainActivity">
 
   <ImageView
-        android:id="@+id/imgLogoMarvel"
-        android:layout_width="264dp"
-        android:layout_height="117dp"
-        android:layout_marginStart="64dp"
-        android:layout_marginLeft="64dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="66dp"
-        android:layout_marginRight="66dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/logo_marvel_time_branco" />
+      android:id="@+id/imgLogoMarvel"
+      android:layout_width="264dp"
+      android:layout_height="117dp"
+      android:layout_marginStart="65dp"
+      android:layout_marginLeft="65dp"
+      android:layout_marginTop="8dp"
+      android:layout_marginEnd="65dp"
+      android:layout_marginRight="65dp"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:srcCompat="@drawable/logo_marvel_time_branco" />
 
     <TextView
         android:id="@+id/textView"
@@ -84,71 +85,12 @@
   <TextView
         android:id="@+id/textViewNome"
         style="@style/titulo"
-        android:text="Olá, Marvel Fã"
+        android:text="@string/ola"
         app:layout_constraintBottom_toBottomOf="@+id/textView"
         app:layout_constraintEnd_toEndOf="@+id/textView"
         app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="@+id/textView"
         app:layout_constraintTop_toTopOf="@+id/textView"
         app:layout_constraintVertical_bias="0.093" />
-
-  <com.michaldrabik.tapbarmenulib.TapBarMenu
-      android:id="@+id/tapBarMenu"
-      android:layout_width="match_parent"
-      android:layout_height="52dp"
-      android:layout_gravity="bottom"
-      android:layout_margin="0dp"
-      android:layout_marginBottom="8dp"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:tbm_backgroundColor="@color/branco"
-      app:tbm_buttonSize="52dp"
-      app:tbm_iconClosed="@drawable/img_menu"
-      app:tbm_iconOpened="@drawable/img_close"
-      app:tbm_menuAnchor="bottom">
-
-    <ImageView
-        android:id="@+id/item1"
-        android:layout_width="0dp"
-        android:layout_height="match_parent"
-        android:layout_weight="1"
-        android:paddingTop="10dp"
-        android:paddingBottom="10dp"
-        android:src="@drawable/img_home"
-        tools:visibility="visible" />
-
-    <ImageView
-        android:id="@+id/item2"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:paddingTop="10dp"
-        android:paddingBottom="10dp"
-        android:src="@drawable/img_favoritos" />
-
-    <Space
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1" />
-
-    <ImageView
-        android:id="@+id/item3"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:paddingTop="10dp"
-        android:paddingBottom="10dp"
-        android:src="@drawable/img_hferro" />
-
-    <ImageView
-        android:id="@+id/item4"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:paddingTop="10dp"
-        android:paddingBottom="10dp"
-        android:src="@drawable/img_quiz_menu" />
-  </com.michaldrabik.tapbarmenulib.TapBarMenu>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
