--- app/src/main/res/layout/activity_menu.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/preto"
-    tools:context=".MenuActivity">
-
-    <FrameLayout
-        android:id="@+id/fragment_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <com.michaldrabik.tapbarmenulib.TapBarMenu
-            android:id="@+id/tapBarMenu"
-            android:layout_width="match_parent"
-            android:layout_height="52dp"
-            app:tbm_buttonSize="52dp"
-            android:layout_gravity="bottom"
-            android:layout_marginBottom="8dp"
-            app:tbm_backgroundColor="@color/branco"
-            app:tbm_iconClosed="@drawable/img_menu"
-            app:tbm_iconOpened="@drawable/img_close"
-            app:tbm_menuAnchor="bottom">
-
-            <ImageView
-                android:id="@+id/item1"
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1"
-                android:paddingTop="10dp"
-                android:paddingBottom="10dp"
-                android:src="@drawable/img_home"
-                tools:visibility="visible"/>
-
-            <ImageView
-                android:id="@+id/item2"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:paddingTop="10dp"
-                android:paddingBottom="10dp"
-                android:src="@drawable/img_favoritos"/>
-
-            <Space
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"/>
-
-            <ImageView
-                android:id="@+id/item3"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:paddingTop="10dp"
-                android:paddingBottom="10dp"
-                android:src="@drawable/img_hferro"/>
-
-            <ImageView
-                android:id="@+id/item4"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:paddingTop="10dp"
-                android:paddingBottom="10dp"
-                android:src="@drawable/img_quiz_menu"/>
-        </com.michaldrabik.tapbarmenulib.TapBarMenu>
- </FrameLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
