+++ app/src/main/res/layout/fragment_favoritos.xml
@@ -20,8 +20,8 @@
 
     <ImageView
         android:id="@+id/logo"
-        android:layout_width="281.32dp"
-        android:layout_height="142.5dp"
+        android:layout_width="261dp"
+        android:layout_height="105dp"
         android:layout_marginStart="64dp"
         android:layout_marginLeft="64dp"
         android:layout_marginTop="8dp"
@@ -67,7 +67,7 @@
         android:layout_height="24dp"
         android:layout_marginStart="16dp"
         android:layout_marginLeft="16dp"
-        android:layout_marginTop="36dp"
+        android:layout_marginTop="44dp"
         android:layout_marginEnd="16dp"
         android:layout_marginRight="16dp"
         android:src="@android:drawable/btn_star_big_on"
@@ -80,7 +80,7 @@
         android:id="@+id/textView"
         android:layout_width="195dp"
         android:layout_height="56dp"
-        android:layout_marginTop="40dp"
+        android:layout_marginTop="32dp"
         android:layout_marginEnd="16dp"
         android:layout_marginRight="16dp"
         android:width="295dp"
@@ -99,7 +99,7 @@
         android:id="@+id/textView3"
         android:layout_width="176dp"
         android:layout_height="47dp"
-        android:layout_marginTop="44dp"
+        android:layout_marginTop="32dp"
         android:width="295dp"
         android:height="57.5dp"
         android:text="@string/doutor_estranho"
@@ -118,7 +118,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginLeft="16dp"
-        android:layout_marginTop="24dp"
+        android:layout_marginTop="32dp"
         android:layout_marginEnd="16dp"
         android:layout_marginRight="16dp"
         android:layout_marginBottom="16dp"
